+++ test/home-assistant-main-mock.js
@@ -3,6 +3,6 @@ import { LitElement, html as litHtml, css as litCss } from "lit-element";
 LitElement.prototype.html = litHtml;
 LitElement.prototype.css = litCss;
 
-class HuiErrorEntityRow extends LitElement {}
+class HomeAssistantMain extends LitElement {}
 
-customElements.define("hui-error-entity-row", HuiErrorEntityRow);
+customElements.define("home-assistant-main", HomeAssistantMain);